fix account tests
